#2 - Patch commit

+++ src/headers/drawer.h
@@ -18,7 +18,7 @@ public:
     ~Drawer();
 
     void lineBrezenham(glm::vec2 from, glm::vec2 to, TGAColor color);
-    void drawTriangle(glm::uvec2 a, glm::uvec2 b, glm::uvec2 c, TGAColor color);
+    void drawTriangle(glm::vec2 a, glm::vec2 b, glm::vec2 c, TGAColor color);
 
     void drawWireframe(Model* m, TGAColor color);
     void drawModel(Model* m, TGAColor color);
